Rollbar.warning instead of Rollbar.error
- Ran into this during an outage we received from 511 endpoints. Chatted
with @darkodosenovic, @jessedoyle, and @mvandenbeuken; if there are
outages on a system we are calling, the calling system should be doing
the alerting. Even in the case of the api.ama.ab.ca it should alert. In
the case of apim.ama.ab.ca and soa.ama.ab.ca we can rely on Loggly
alerts to go to the appropriate team

- TODO: we might have to update the Loggly alerts for api.ama.ab.ca to
go to Red team on-call when there are outages/errors, but Rollbar should
currently be handling this.

- Bumps version number: I beleive a [patch is semantically
correct](https://semver.org) but it seems like a significant change, so
perhaps we need to bump further?

See:
- https://github.com/amaabca/sentinel_web/pull/1325 for further
information